renaming stimuli in psychopy render to displays
